backend/requirements.txt
openai==1.2.0
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0
flask-cors==4.0.0
